app_new/requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
torch>=2.0.0
transformers>=4.30.0
huggingface-hub>=0.19.0
python-dotenv>=1.0.0 